Final Move and Import Fixing
Ready for pull request
